// renderer/app/(menu)/menu/_types/types.ts
export type FeatureCard = {
  icon: React.ReactNode
  title: string
  description: string
  href: string
  color: string
  bg: string
  border: string
  disabled?: boolean
}

export type AdminCard = {
  icon: React.ReactNode
  title: string
  description: string
  href: string
  disabled?: boolean
  color: string
  bg: string
}